Move range parsing to separate file
